refactor and code reorganise
