+++ QuestEssentials/Messages/SellItemMessage.cs
@@ -1,10 +1,5 @@
-﻿using QuestEssentials.Quests.Story;
-using StardewValley;
+﻿using StardewValley;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace QuestEssentials.Messages
 {
