+++ QuestEssentials/QuestEssentials.csproj
@@ -20,4 +20,8 @@
     <Reference Include="QuestFramework" HintPath="$(GamePath)\Mods\QuestFramework\QuestFramework.dll" Private="false" Condition="!Exists('$(SolutionDir)\QuestFramework\QuestFramework.csproj')" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Quests\Story\" />
+  </ItemGroup>
+
 </Project>
