+++ QuestEssentials/Tasks/BasicTask.cs
@@ -1,11 +1,6 @@
-﻿using Newtonsoft.Json.Linq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using QuestEssentials.Messages;
 
-namespace QuestEssentials.Quests.Story.Tasks
+namespace QuestEssentials.Tasks
 {
     class BasicTask : StoryQuestTask
     {
